Made project structure more consistent
