+++ Client/src/Models/TaskEnums.ts
@@ -1,0 +1,4 @@
+export type StatusEnum = 0 | 1 | 2 | 3;
+export type EstimateEnum = 0 | 1 | 2;
+export type TypeEnum = 0 | 1 | 2;
+export type PriorityEnum = 0 | 1 | 2;
